Shopping Cart with Strits Framework,Beans,Tag libraries & EL
This is the final project for the course WWW Technologies which takes
place in University of Thessaly.
This web app is a shopping cart. In the first page(index.jsp) the
client chooses the currency that he/she desires. After that he is
forwarded to the shopping cart where the prices of the cart are
modified according to the chosen currency. the prices of the items are
in the file pricelist.xml. I have use my own tag library that does the
parsing of the file so that the code of the parser can be reusable. All
the prices, the partial costs and the sum of all costs are converted to
the chosen currency ( rates are taken by google so they are real and
dynamic ).